PurchaserADD

+++ src/main/java/com/mundomuebles/mundo_muebles_fd/application/PurchaserAble.java
@@ -2,11 +2,12 @@ package com.mundomuebles.mundo_muebles_fd.application;
 
 import com.mundomuebles.mundo_muebles_fd.domain.PurchaserDTO;
 
+import java.math.BigInteger;
 import java.util.List;
 
 public interface PurchaserAble {
     PurchaserDTO save(PurchaserDTO purchaserDTO);
     PurchaserDTO update(PurchaserDTO purchaserDTO);
-    boolean delete (Integer id);
+    boolean delete (BigInteger id);
     List<PurchaserDTO> list();
 }
