+++ src/main/java/com/mundomuebles/mundo_muebles_fd/application/PurchaserImpl.java
@@ -5,6 +5,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
 
+import java.math.BigInteger;
 import java.util.List;
 
 
@@ -27,7 +28,7 @@ public class PurchaserImpl implements PurchaserAble {
     }
 
     @Override
-    public boolean delete(Integer id) {
+    public boolean delete(BigInteger id) {
         return false;
     }
