+++ src/main/java/com/mundomuebles/mundo_muebles_fd/infrastructure/adapters/PostgresPurchaserRepository.java
@@ -8,6 +8,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigInteger;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -29,7 +30,7 @@ public class PostgresPurchaserRepository implements PurchaserAble {
     }
 
     @Override
-    public boolean delete(Integer id) {
+    public boolean delete(BigInteger id) {
         purchaserRepository.deleteById(id);
         return true;
     }
