+++ src/main/java/com/mundomuebles/mundo_muebles_fd/infrastructure/repositories/PurchaserRepository.java
@@ -2,5 +2,7 @@ package com.mundomuebles.mundo_muebles_fd.infrastructure.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PurchaserRepository extends JpaRepository <Purchaser, Integer> {
+import java.math.BigInteger;
+
+public interface PurchaserRepository extends JpaRepository <Purchaser, BigInteger> {
 }
